Move jwt methodes to shared.
