+++ apps/server/src/shared/common/utils/index.ts
@@ -1,3 +1,4 @@
 export { GuardAgainst } from './guard-against';
 export { SortHelper } from './sort-helper';
 export { getResolvedValues, isFulfilled } from './promise';
+export { extractJwtFromHeader, JwtExtractor } from './jwt';
